move json serializer outside of request handler to allow easy working with serializer in certain scenarios
