+++ src/Http/Client/Ocluse.LiquidSnow.Http.Client.csproj
@@ -7,7 +7,7 @@
     <Nullable>enable</Nullable>
 	<GeneratePackageOnBuild>True</GeneratePackageOnBuild>
 	<GenerateDocumentationFile>true</GenerateDocumentationFile>
-	<Version>0.5.2-preview1.0</Version>
+	<Version>0.5.2-preview2.0</Version>
 	<Authors>Ocluse</Authors>
 	<PackageOutputPath>..\..\..\build</PackageOutputPath>
 	<Description>Provides utilities for implementing common HTTP client requests for CRUD patterns.</Description>
